Move BatchRowCount tests
